requirements.txt
# Core ML libraries
torch>=1.13.0
transformers>=4.28.1,<5.0.0
trl>=0.7.4
peft>=0.4.0
accelerate>=0.20.0
datasets>=2.0.0

# API and external services
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
boto3>=1.26.0
tenacity>=8.0.0

# Utilities
ml_collections
gradio>=4.0.0
gradio-client>=0.7.0
fschat>=0.2.20
matplotlib>=3.5.0
sentencepiece>=0.1.99
tqdm>=4.64.0
pandas>=1.5.0
numpy>=1.21.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Local packages (install in editable mode)
-e ./peft
-e ./just_eval
